Moving helper function file out of bin/
