+++ src/bin/main.ts
@@ -4,10 +4,10 @@ import Slack from "../lib/io/slack";
 import log from '../lib/log/logger';
 import { Database } from "../lib/model/database";
 import { cli } from "../lib/parameters/cli-args";
+import { ioFromCliArgs } from '../lib/parameters/cli-io';
 import { envConfig, runLoopConfigFromENV, slackConfigFromENV } from "../lib/parameters/env-config";
 import { AttachableError, KnownError } from "../lib/util/errors";
 import run from "../lib/util/runner";
-import { ioFromCliArgs } from './cli-args';
 
 main();
 async function main() {
